rename tick files to tick-mark
