+++ projects/ngx-d3timeline/src/lib/axis/axis.ts
@@ -1,4 +1,4 @@
-import { TickMark } from './tick';
+import { TickMark } from './tick-mark';
 import { Line } from './line';
 
 export interface Axis {
